requirements.txt
airtable-python-wrapper==0.15.2
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==2.10
openpyxl==3.0.7
pkg-resources==0.0.0
requests==2.25.1
schedule==1.1.0
selenium==3.141.0
urllib3==1.26.4
